Revert "Update phone number on frontend after number is set (#2256)"
This reverts commit a5d52c7cc370b05a442da9bd307845f4ad94da29.
