+++ api-js/src/user/mutations/__tests__/set-phone-number.test.js
@@ -8,7 +8,7 @@ import frenchMessages from '../../../locale/fr/messages'
 import { databaseOptions } from '../../../../database-options'
 import { createQuerySchema } from '../../../query'
 import { createMutationSchema } from '../../../mutation'
-import { cleanseInput, decryptPhoneNumber } from '../../../validators'
+import { cleanseInput } from '../../../validators'
 import { tokenize, userRequired } from '../../../auth'
 import { loadUserByUserName, loadUserByKey } from '../../loaders'
 
@@ -85,19 +85,15 @@ describe('user sets a new phone number', () => {
             emailValidated: false,
           })
         })
-        it('returns status text and updated user', async () => {
-          const newPhoneNumber = '+12345678901'
+        it('returns status text', async () => {
           const response = await graphql(
             schema,
             `
               mutation {
-                setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                   result {
                     ... on SetPhoneNumberResult {
                       status
-                      user {
-                        phoneNumber
-                      }
                     }
                     ... on SetPhoneNumberError {
                       code
@@ -123,7 +119,6 @@ describe('user sets a new phone number', () => {
               },
               validators: {
                 cleanseInput,
-                decryptPhoneNumber,
               },
               loaders: {
                 loadUserByKey: loadUserByKey({ query }),
@@ -140,9 +135,6 @@ describe('user sets a new phone number', () => {
                 result: {
                   status:
                     'Phone number has been successfully set, you will receive a verification text message shortly.',
-                  user: {
-                    phoneNumber: newPhoneNumber,
-                  },
                 },
               },
             },
@@ -156,13 +148,12 @@ describe('user sets a new phone number', () => {
 
           expect(response).toEqual(expectedResult)
           expect(mockNotify).toHaveBeenCalledWith({
-            phoneNumber: newPhoneNumber,
+            phoneNumber: '+12345678901',
             user,
           })
           expect(consoleOutput).toEqual([
             `User: ${user._key} successfully set phone number.`,
           ])
-          expect(decryptPhoneNumber(user.phoneDetails)).toEqual(newPhoneNumber)
         })
       })
       describe('user is phone validated', () => {
@@ -179,19 +170,15 @@ describe('user sets a new phone number', () => {
               tfaSendMethod: 'none',
             })
           })
-          it('returns status text and updated user', async () => {
-            const newPhoneNumber = '+12345678901'
+          it('returns status text', async () => {
             const response = await graphql(
               schema,
               `
                 mutation {
-                  setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                  setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                     result {
                       ... on SetPhoneNumberResult {
                         status
-                        user {
-                          phoneNumber
-                        }
                       }
                       ... on SetPhoneNumberError {
                         code
@@ -217,7 +204,6 @@ describe('user sets a new phone number', () => {
                 },
                 validators: {
                   cleanseInput,
-                  decryptPhoneNumber,
                 },
                 loaders: {
                   loadUserByKey: loadUserByKey({ query }),
@@ -234,9 +220,6 @@ describe('user sets a new phone number', () => {
                   result: {
                     status:
                       'Phone number has been successfully set, you will receive a verification text message shortly.',
-                    user: {
-                      phoneNumber: newPhoneNumber,
-                    },
                   },
                 },
               },
@@ -250,29 +233,22 @@ describe('user sets a new phone number', () => {
 
             expect(response).toEqual(expectedResult)
             expect(mockNotify).toHaveBeenCalledWith({
-              phoneNumber: newPhoneNumber,
+              phoneNumber: '+12345678901',
               user,
             })
             expect(consoleOutput).toEqual([
               `User: ${user._key} successfully set phone number.`,
             ])
-            expect(decryptPhoneNumber(user.phoneDetails)).toEqual(
-              newPhoneNumber,
-            )
           })
           it('tfaSendMethod stays as none', async () => {
-            const newPhoneNumber = '+12345678901'
             await graphql(
               schema,
               `
                 mutation {
-                  setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                  setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                     result {
                       ... on SetPhoneNumberResult {
                         status
-                        user {
-                          phoneNumber
-                        }
                       }
                       ... on SetPhoneNumberError {
                         code
@@ -298,7 +274,6 @@ describe('user sets a new phone number', () => {
                 },
                 validators: {
                   cleanseInput,
-                  decryptPhoneNumber,
                 },
                 loaders: {
                   loadUserByKey: loadUserByKey({ query }),
@@ -316,18 +291,14 @@ describe('user sets a new phone number', () => {
             expect(user.tfaSendMethod).toEqual('none')
           })
           it('sets phoneValidated to false', async () => {
-            const newPhoneNumber = '+12345678901'
             await graphql(
               schema,
               `
                 mutation {
-                  setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                  setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                     result {
                       ... on SetPhoneNumberResult {
                         status
-                        user {
-                          phoneNumber
-                        }
                       }
                       ... on SetPhoneNumberError {
                         code
@@ -353,7 +324,6 @@ describe('user sets a new phone number', () => {
                 },
                 validators: {
                   cleanseInput,
-                  decryptPhoneNumber,
                 },
                 loaders: {
                   loadUserByKey: loadUserByKey({ query }),
@@ -384,19 +354,15 @@ describe('user sets a new phone number', () => {
               tfaSendMethod: 'email',
             })
           })
-          it('returns status text and update user', async () => {
-            const newPhoneNumber = '+12345678901'
+          it('returns status text', async () => {
             const response = await graphql(
               schema,
               `
                 mutation {
-                  setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                  setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                     result {
                       ... on SetPhoneNumberResult {
                         status
-                        user {
-                          phoneNumber
-                        }
                       }
                       ... on SetPhoneNumberError {
                         code
@@ -422,7 +388,6 @@ describe('user sets a new phone number', () => {
                 },
                 validators: {
                   cleanseInput,
-                  decryptPhoneNumber,
                 },
                 loaders: {
                   loadUserByKey: loadUserByKey({ query }),
@@ -439,9 +404,6 @@ describe('user sets a new phone number', () => {
                   result: {
                     status:
                       'Phone number has been successfully set, you will receive a verification text message shortly.',
-                    user: {
-                      phoneNumber: newPhoneNumber,
-                    },
                   },
                 },
               },
@@ -455,29 +417,22 @@ describe('user sets a new phone number', () => {
 
             expect(response).toEqual(expectedResult)
             expect(mockNotify).toHaveBeenCalledWith({
-              phoneNumber: newPhoneNumber,
+              phoneNumber: '+12345678901',
               user,
             })
             expect(consoleOutput).toEqual([
               `User: ${user._key} successfully set phone number.`,
             ])
-            expect(decryptPhoneNumber(user.phoneDetails)).toEqual(
-              newPhoneNumber,
-            )
           })
           it('tfaSendMethod stays as email', async () => {
-            const newPhoneNumber = '+12345678901'
             await graphql(
               schema,
               `
                 mutation {
-                  setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                  setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                     result {
                       ... on SetPhoneNumberResult {
                         status
-                        user {
-                          phoneNumber
-                        }
                       }
                       ... on SetPhoneNumberError {
                         code
@@ -503,7 +458,6 @@ describe('user sets a new phone number', () => {
                 },
                 validators: {
                   cleanseInput,
-                  decryptPhoneNumber,
                 },
                 loaders: {
                   loadUserByKey: loadUserByKey({ query }),
@@ -521,18 +475,14 @@ describe('user sets a new phone number', () => {
             expect(user.tfaSendMethod).toEqual('email')
           })
           it('sets phoneValidated to false', async () => {
-            const newPhoneNumber = '+12345678901'
             await graphql(
               schema,
               `
                 mutation {
-                  setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                  setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                     result {
                       ... on SetPhoneNumberResult {
                         status
-                        user {
-                          phoneNumber
-                        }
                       }
                       ... on SetPhoneNumberError {
                         code
@@ -558,7 +508,6 @@ describe('user sets a new phone number', () => {
                 },
                 validators: {
                   cleanseInput,
-                  decryptPhoneNumber,
                 },
                 loaders: {
                   loadUserByKey: loadUserByKey({ query }),
@@ -589,19 +538,15 @@ describe('user sets a new phone number', () => {
               tfaSendMethod: 'email',
             })
           })
-          it('returns status text and updated user', async () => {
-            const newPhoneNumber = '+12345678901'
+          it('returns status text', async () => {
             const response = await graphql(
               schema,
               `
                 mutation {
-                  setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                  setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                     result {
                       ... on SetPhoneNumberResult {
                         status
-                        user {
-                          phoneNumber
-                        }
                       }
                       ... on SetPhoneNumberError {
                         code
@@ -627,7 +572,6 @@ describe('user sets a new phone number', () => {
                 },
                 validators: {
                   cleanseInput,
-                  decryptPhoneNumber,
                 },
                 loaders: {
                   loadUserByKey: loadUserByKey({ query }),
@@ -644,9 +588,6 @@ describe('user sets a new phone number', () => {
                   result: {
                     status:
                       'Phone number has been successfully set, you will receive a verification text message shortly.',
-                    user: {
-                      phoneNumber: newPhoneNumber,
-                    },
                   },
                 },
               },
@@ -660,29 +601,22 @@ describe('user sets a new phone number', () => {
 
             expect(response).toEqual(expectedResult)
             expect(mockNotify).toHaveBeenCalledWith({
-              phoneNumber: newPhoneNumber,
+              phoneNumber: '+12345678901',
               user,
             })
             expect(consoleOutput).toEqual([
               `User: ${user._key} successfully set phone number.`,
             ])
-            expect(decryptPhoneNumber(user.phoneDetails)).toEqual(
-              newPhoneNumber,
-            )
           })
           it('sets tfaSendMethod to email', async () => {
-            const newPhoneNumber = '+12345678901'
             await graphql(
               schema,
               `
                 mutation {
-                  setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                  setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                     result {
                       ... on SetPhoneNumberResult {
                         status
-                        user {
-                          phoneNumber
-                        }
                       }
                       ... on SetPhoneNumberError {
                         code
@@ -708,7 +642,6 @@ describe('user sets a new phone number', () => {
                 },
                 validators: {
                   cleanseInput,
-                  decryptPhoneNumber,
                 },
                 loaders: {
                   loadUserByKey: loadUserByKey({ query }),
@@ -726,18 +659,14 @@ describe('user sets a new phone number', () => {
             expect(user.tfaSendMethod).toEqual('email')
           })
           it('sets phoneValidated to false', async () => {
-            const newPhoneNumber = '+12345678901'
             await graphql(
               schema,
               `
                 mutation {
-                  setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                  setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                     result {
                       ... on SetPhoneNumberResult {
                         status
-                        user {
-                          phoneNumber
-                        }
                       }
                       ... on SetPhoneNumberError {
                         code
@@ -763,7 +692,6 @@ describe('user sets a new phone number', () => {
                 },
                 validators: {
                   cleanseInput,
-                  decryptPhoneNumber,
                 },
                 loaders: {
                   loadUserByKey: loadUserByKey({ query }),
@@ -802,19 +730,14 @@ describe('user sets a new phone number', () => {
             .fn()
             .mockRejectedValue(new Error('Database error occurred.'))
 
-          const newPhoneNumber = '+12345678901'
-
           const response = await graphql(
             schema,
             `
               mutation {
-                setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                   result {
                     ... on SetPhoneNumberResult {
                       status
-                      user {
-                        phoneNumber
-                      }
                     }
                     ... on SetPhoneNumberError {
                       code
@@ -840,7 +763,6 @@ describe('user sets a new phone number', () => {
               },
               validators: {
                 cleanseInput,
-                decryptPhoneNumber,
               },
               loaders: {
                 loadUserByKey: loaderById,
@@ -890,19 +812,15 @@ describe('user sets a new phone number', () => {
             emailValidated: false,
           })
         })
-        it('returns status text and updated user', async () => {
-          const newPhoneNumber = '+12345678901'
+        it('returns status text', async () => {
           const response = await graphql(
             schema,
             `
               mutation {
-                setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                   result {
                     ... on SetPhoneNumberResult {
                       status
-                      user {
-                        phoneNumber
-                      }
                     }
                     ... on SetPhoneNumberError {
                       code
@@ -928,7 +846,6 @@ describe('user sets a new phone number', () => {
               },
               validators: {
                 cleanseInput,
-                decryptPhoneNumber,
               },
               loaders: {
                 loadUserByKey: loadUserByKey({ query }),
@@ -944,9 +861,6 @@ describe('user sets a new phone number', () => {
               setPhoneNumber: {
                 result: {
                   status: 'todo',
-                  user: {
-                    phoneNumber: newPhoneNumber,
-                  },
                 },
               },
             },
@@ -960,13 +874,12 @@ describe('user sets a new phone number', () => {
 
           expect(response).toEqual(expectedResult)
           expect(mockNotify).toHaveBeenCalledWith({
-            phoneNumber: newPhoneNumber,
+            phoneNumber: '+12345678901',
             user,
           })
           expect(consoleOutput).toEqual([
             `User: ${user._key} successfully set phone number.`,
           ])
-          expect(decryptPhoneNumber(user.phoneDetails)).toEqual(newPhoneNumber)
         })
       })
       describe('user is phone validated', () => {
@@ -983,19 +896,15 @@ describe('user sets a new phone number', () => {
               tfaSendMethod: 'none',
             })
           })
-          it('returns status text and updated user', async () => {
-            const newPhoneNumber = '+12345678901'
+          it('returns status text', async () => {
             const response = await graphql(
               schema,
               `
                 mutation {
-                  setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                  setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                     result {
                       ... on SetPhoneNumberResult {
                         status
-                        user {
-                          phoneNumber
-                        }
                       }
                       ... on SetPhoneNumberError {
                         code
@@ -1021,7 +930,6 @@ describe('user sets a new phone number', () => {
                 },
                 validators: {
                   cleanseInput,
-                  decryptPhoneNumber,
                 },
                 loaders: {
                   loadUserByKey: loadUserByKey({ query }),
@@ -1037,9 +945,6 @@ describe('user sets a new phone number', () => {
                 setPhoneNumber: {
                   result: {
                     status: 'todo',
-                    user: {
-                      phoneNumber: newPhoneNumber,
-                    },
                   },
                 },
               },
@@ -1053,29 +958,22 @@ describe('user sets a new phone number', () => {
 
             expect(response).toEqual(expectedResult)
             expect(mockNotify).toHaveBeenCalledWith({
-              phoneNumber: newPhoneNumber,
+              phoneNumber: '+12345678901',
               user,
             })
             expect(consoleOutput).toEqual([
               `User: ${user._key} successfully set phone number.`,
             ])
-            expect(decryptPhoneNumber(user.phoneDetails)).toEqual(
-              newPhoneNumber,
-            )
           })
           it('tfaSendMethod stays as none', async () => {
-            const newPhoneNumber = '+12345678901'
             await graphql(
               schema,
               `
                 mutation {
-                  setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                  setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                     result {
                       ... on SetPhoneNumberResult {
                         status
-                        user {
-                          phoneNumber
-                        }
                       }
                       ... on SetPhoneNumberError {
                         code
@@ -1101,7 +999,6 @@ describe('user sets a new phone number', () => {
                 },
                 validators: {
                   cleanseInput,
-                  decryptPhoneNumber,
                 },
                 loaders: {
                   loadUserByKey: loadUserByKey({ query }),
@@ -1119,18 +1016,14 @@ describe('user sets a new phone number', () => {
             expect(user.tfaSendMethod).toEqual('none')
           })
           it('sets phoneValidated to false', async () => {
-            const newPhoneNumber = '+12345678901'
             await graphql(
               schema,
               `
                 mutation {
-                  setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                  setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                     result {
                       ... on SetPhoneNumberResult {
                         status
-                        user {
-                          phoneNumber
-                        }
                       }
                       ... on SetPhoneNumberError {
                         code
@@ -1156,7 +1049,6 @@ describe('user sets a new phone number', () => {
                 },
                 validators: {
                   cleanseInput,
-                  decryptPhoneNumber,
                 },
                 loaders: {
                   loadUserByKey: loadUserByKey({ query }),
@@ -1188,18 +1080,14 @@ describe('user sets a new phone number', () => {
             })
           })
           it('returns status text', async () => {
-            const newPhoneNumber = '+12345678901'
             const response = await graphql(
               schema,
               `
                 mutation {
-                  setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                  setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                     result {
                       ... on SetPhoneNumberResult {
                         status
-                        user {
-                          phoneNumber
-                        }
                       }
                       ... on SetPhoneNumberError {
                         code
@@ -1225,7 +1113,6 @@ describe('user sets a new phone number', () => {
                 },
                 validators: {
                   cleanseInput,
-                  decryptPhoneNumber,
                 },
                 loaders: {
                   loadUserByKey: loadUserByKey({ query }),
@@ -1241,9 +1128,6 @@ describe('user sets a new phone number', () => {
                 setPhoneNumber: {
                   result: {
                     status: 'todo',
-                    user: {
-                      phoneNumber: newPhoneNumber,
-                    },
                   },
                 },
               },
@@ -1257,29 +1141,22 @@ describe('user sets a new phone number', () => {
 
             expect(response).toEqual(expectedResult)
             expect(mockNotify).toHaveBeenCalledWith({
-              phoneNumber: newPhoneNumber,
+              phoneNumber: '+12345678901',
               user,
             })
             expect(consoleOutput).toEqual([
               `User: ${user._key} successfully set phone number.`,
             ])
-            expect(decryptPhoneNumber(user.phoneDetails)).toEqual(
-              newPhoneNumber,
-            )
           })
           it('tfaSendMethod stays as email', async () => {
-            const newPhoneNumber = '+12345678901'
             await graphql(
               schema,
               `
                 mutation {
-                  setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                  setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                     result {
                       ... on SetPhoneNumberResult {
                         status
-                        user {
-                          phoneNumber
-                        }
                       }
                       ... on SetPhoneNumberError {
                         code
@@ -1305,7 +1182,6 @@ describe('user sets a new phone number', () => {
                 },
                 validators: {
                   cleanseInput,
-                  decryptPhoneNumber,
                 },
                 loaders: {
                   loadUserByKey: loadUserByKey({ query }),
@@ -1323,18 +1199,14 @@ describe('user sets a new phone number', () => {
             expect(user.tfaSendMethod).toEqual('email')
           })
           it('sets phoneValidated to false', async () => {
-            const newPhoneNumber = '+12345678901'
             await graphql(
               schema,
               `
                 mutation {
-                  setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                  setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                     result {
                       ... on SetPhoneNumberResult {
                         status
-                        user {
-                          phoneNumber
-                        }
                       }
                       ... on SetPhoneNumberError {
                         code
@@ -1360,7 +1232,6 @@ describe('user sets a new phone number', () => {
                 },
                 validators: {
                   cleanseInput,
-                  decryptPhoneNumber,
                 },
                 loaders: {
                   loadUserByKey: loadUserByKey({ query }),
@@ -1392,18 +1263,14 @@ describe('user sets a new phone number', () => {
             })
           })
           it('returns status text', async () => {
-            const newPhoneNumber = '+12345678901'
             const response = await graphql(
               schema,
               `
                 mutation {
-                  setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                  setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                     result {
                       ... on SetPhoneNumberResult {
                         status
-                        user {
-                          phoneNumber
-                        }
                       }
                       ... on SetPhoneNumberError {
                         code
@@ -1429,7 +1296,6 @@ describe('user sets a new phone number', () => {
                 },
                 validators: {
                   cleanseInput,
-                  decryptPhoneNumber,
                 },
                 loaders: {
                   loadUserByKey: loadUserByKey({ query }),
@@ -1445,9 +1311,6 @@ describe('user sets a new phone number', () => {
                 setPhoneNumber: {
                   result: {
                     status: 'todo',
-                    user: {
-                      phoneNumber: newPhoneNumber,
-                    },
                   },
                 },
               },
@@ -1461,29 +1324,22 @@ describe('user sets a new phone number', () => {
 
             expect(response).toEqual(expectedResult)
             expect(mockNotify).toHaveBeenCalledWith({
-              phoneNumber: newPhoneNumber,
+              phoneNumber: '+12345678901',
               user,
             })
             expect(consoleOutput).toEqual([
               `User: ${user._key} successfully set phone number.`,
             ])
-            expect(decryptPhoneNumber(user.phoneDetails)).toEqual(
-              newPhoneNumber,
-            )
           })
           it('sets tfaSendMethod to email', async () => {
-            const newPhoneNumber = '+12345678901'
             await graphql(
               schema,
               `
                 mutation {
-                  setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                  setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                     result {
                       ... on SetPhoneNumberResult {
                         status
-                        user {
-                          phoneNumber
-                        }
                       }
                       ... on SetPhoneNumberError {
                         code
@@ -1509,7 +1365,6 @@ describe('user sets a new phone number', () => {
                 },
                 validators: {
                   cleanseInput,
-                  decryptPhoneNumber,
                 },
                 loaders: {
                   loadUserByKey: loadUserByKey({ query }),
@@ -1527,18 +1382,14 @@ describe('user sets a new phone number', () => {
             expect(user.tfaSendMethod).toEqual('email')
           })
           it('sets phoneValidated to false', async () => {
-            const newPhoneNumber = '+12345678901'
             await graphql(
               schema,
               `
                 mutation {
-                  setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                  setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                     result {
                       ... on SetPhoneNumberResult {
                         status
-                        user {
-                          phoneNumber
-                        }
                       }
                       ... on SetPhoneNumberError {
                         code
@@ -1564,7 +1415,6 @@ describe('user sets a new phone number', () => {
                 },
                 validators: {
                   cleanseInput,
-                  decryptPhoneNumber,
                 },
                 loaders: {
                   loadUserByKey: loadUserByKey({ query }),
@@ -1603,13 +1453,11 @@ describe('user sets a new phone number', () => {
             .fn()
             .mockRejectedValue(new Error('Database error occurred.'))
 
-          const newPhoneNumber = '+12345678901'
-
           const response = await graphql(
             schema,
             `
               mutation {
-                setPhoneNumber(input: { phoneNumber: "${newPhoneNumber}" }) {
+                setPhoneNumber(input: { phoneNumber: "+12345678901" }) {
                   result {
                     ... on SetPhoneNumberResult {
                       status
