+++ api-js/src/user/objects/__tests__/set-phone-number-result.test.js
@@ -1,6 +1,6 @@
 import { GraphQLString } from 'graphql'
 
-import { setPhoneNumberResultType, userPersonalType } from '../index'
+import { setPhoneNumberResultType } from '../index'
 
 describe('given the setPhoneNumberResultType object', () => {
   describe('testing the field definitions', () => {
@@ -10,12 +10,6 @@ describe('given the setPhoneNumberResultType object', () => {
       expect(demoType).toHaveProperty('status')
       expect(demoType.status.type).toMatchObject(GraphQLString)
     })
-    it('has a user field', () => {
-      const demoType = setPhoneNumberResultType.getFields()
-
-      expect(demoType).toHaveProperty('user')
-      expect(demoType.user.type).toMatchObject(userPersonalType)
-    })
   })
 
   describe('testing the field resolvers', () => {
@@ -24,19 +18,6 @@ describe('given the setPhoneNumberResultType object', () => {
         const demoType = setPhoneNumberResultType.getFields()
 
         expect(demoType.status.resolve({ status: 'status' })).toEqual('status')
-      })
-    })
-    describe('testing the user resolver', () => {
-      it('returns the resolved field', () => {
-        const demoType = setPhoneNumberResultType.getFields()
-
-        const user = {
-          displayName: 'John Doe',
-        }
-
-        expect(demoType.user.resolve({ user })).toEqual({
-          displayName: 'John Doe',
-        })
       })
     })
   })
