+++ api-js/src/user/objects/__tests__/user-personal.test.js
@@ -11,7 +11,6 @@ import { GraphQLEmailAddress, GraphQLPhoneNumber } from 'graphql-scalars'
 import { affiliationConnection } from '../../../affiliation/objects'
 import { userPersonalType } from '../index'
 import { LanguageEnums, TfaSendMethodEnum } from '../../../enums'
-import { decryptPhoneNumber } from '../../../validators'
 
 const { CIPHER_KEY } = process.env
 
@@ -110,13 +109,9 @@ describe('given the user object', () => {
           const phoneDetails = undefined
 
           expect(
-            demoType.phoneNumber.resolve(
-              {
-                phoneDetails,
-              },
-              {},
-              { validators: { decryptPhoneNumber } },
-            ),
+            demoType.phoneNumber.resolve({
+              phoneDetails,
+            }),
           ).toEqual(null)
         })
       })
@@ -127,13 +122,9 @@ describe('given the user object', () => {
           const phoneDetails = null
 
           expect(
-            demoType.phoneNumber.resolve(
-              {
-                phoneDetails,
-              },
-              {},
-              { validators: { decryptPhoneNumber } },
-            ),
+            demoType.phoneNumber.resolve({
+              phoneDetails,
+            }),
           ).toEqual(null)
         })
       })
@@ -158,17 +149,13 @@ describe('given the user object', () => {
           encrypted += cipher.final('hex')
 
           expect(
-            demoType.phoneNumber.resolve(
-              {
-                phoneDetails: {
-                  phoneNumber: encrypted,
-                  iv: phoneDetails.iv,
-                  tag: cipher.getAuthTag().toString('hex'),
-                },
+            demoType.phoneNumber.resolve({
+              phoneDetails: {
+                phoneNumber: encrypted,
+                iv: phoneDetails.iv,
+                tag: cipher.getAuthTag().toString('hex'),
               },
-              {},
-              { validators: { decryptPhoneNumber } },
-            ),
+            }),
           ).toEqual(phoneDetails.phoneNumber)
         })
       })
