+++ api-js/src/user/objects/set-phone-number-result.js
@@ -1,5 +1,4 @@
 import { GraphQLObjectType, GraphQLString } from 'graphql'
-import { userPersonalType } from './user-personal'
 
 export const setPhoneNumberResultType = new GraphQLObjectType({
   name: 'SetPhoneNumberResult',
@@ -11,11 +10,6 @@ export const setPhoneNumberResultType = new GraphQLObjectType({
       description:
         'Informs the user if their phone code was successfully sent.',
       resolve: ({ status }) => status,
-    },
-    user: {
-      type: userPersonalType,
-      description: 'The user who set their phone number.',
-      resolve: ({ user }) => user,
     },
   }),
 })
