+++ api-js/src/validators/index.js
@@ -1,3 +1,2 @@
 export * from './cleanse-input'
-export * from './decrypt-phone-number'
 export * from './slugify'
