+++ frontend/src/graphql/mutations.js
@@ -322,10 +322,6 @@ export const SET_PHONE_NUMBER = gql`
       result {
         ... on SetPhoneNumberResult {
           status
-          user {
-            id
-            phoneNumber
-          }
         }
         ... on SetPhoneNumberError {
           code
@@ -342,10 +338,6 @@ export const VERIFY_PHONE_NUMBER = gql`
       result {
         ... on VerifyPhoneNumberResult {
           status
-          user {
-            id
-            phoneNumber
-          }
         }
         ... on VerifyPhoneNumberError {
           code
